+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased] - ReleaseDate
+### Added
+- `DivBuf`s and `DivBufMut`s now share ownership of the data, so they can live
+  even after the original `DivBufShared` has been dropped.
+  https://github.com/asomers/divbuf/pull/1
+
 ## [0.2.0] - 2018-07-01
 ### Added
 - Implemented `Borrow` and `BorrowMut` for `DivBuf` and `DivBufMut`